Refactor landing location/orientation generation
For BoX, the position and orientation of the Land command is ignored by
the AI if the command is near an airfield, but is used to display in the
player GUI the landing waypoint. This means that currently, the approach
path and landing waypoint point in the opposite direction to the AI
approach path.

Refactor the landing location/orientation determination into the
Airfield object, and for BoX use the runway start point as the landing
location, to match the AI behaviour (also correctly means that takeoff
and landing directions are the same). RoF should not be affected
(although the approach waypoint will move back somewhat, as it will be
offset by both the landing distance and the approach distance).