AAR log parsing update
Work around an issue where kill events would be listed in the AAR for
"BotPilot_XXX". This is caused when a bot bails out of their plane at
too low an altitude and is killed on impact with the ground. The AType12
event for the bot is generated after the bot has bailed out, so the
parent link to the plane is absent. This change parses the bailout
events in order to associate the bot with the original plane.