Identify gunner kills in AAR
Currently planes shot down by gunners are listed as crashed. This patch
updates the parsing to identify these as gunner kills. These are listed
in the AAR debrief, but are not attributed to the gunner's pilot, and do
not appear in the campaign log even if the gunner belongs to the player
squadron, as there's no crew member to attribute them to.