Enable logging of hit/damage events for all planes
The built-in career mode missions have log messages for all planes in
the player's flight, not just kill events. This is achieved by having a
non-printing character at the start of the plane name - the career
missions use 0x01 for plane 1 in the flight, 0x02 for plane 2 etc. When
this is done, the planes are treated as if they were players, with a
fake GUID containing the special hex value. From testing, there don't
seem to be any issues with using 0x01 for all planes, so we do this to
get log messages for all flights.