AAR victory assignment fixes
Recently some reports have come in of claims being denied. Looking into
this, it seems there are two main causes:
 - In 3.006 of BoX, it seems far more common that kill events in the
   logs are uncredited, even if the kill is credited to the player in
   game.
 - The earlier update to add a special character to plane names in order
   to get more logging information means that more damage events are
   reported for all planes. This is confusing the
   crossed-path-with-player logic and causing fuzzy claims to be denied.
Additionally, the phase 2 fuzzy victory assignment logic seems not to
have been working, as it is called before the victory list is populated,
so runs on an empty list.

This change:
 - Removes the assumption that damage events in the logs must have
   either victim or victor as the/a player, and instead tracks damage
   for each victim/victor pair.
 - Updates the crossed-path-with-player logic to use this, which will
   now identify based on damage events for any plane in the player
   flight, not just the player ones
 - Makes the fuzzy victory evaluator actually run
 - Removes the phase 2 random victory assignment (which was never run
   before), as this is now covered by phase 3
 - Updates the fuzzy by damage analysis to handle damage by multiple
   planes, and award the victory to the plane causing the most total
   damage

I think the fuzzy by damage analysis would actually be better handled in
phase 3 as part of claim resolution, but that would be a lot more work
and this seems to work pretty well as a fix in the meantime.