Allow buildings as ground victories
Blocks near airfields are set to the country of the airfield (gives
soldiers walking around). Otherwise blocks behind lines are set to a
country for that side, and blocks close to or between the lines are left
neutral.

Durability of buildings is set based on latest guidelines from devs

Fix airfield static vehicles - these were being output as Vehicles
rather than Blocks, so weren't appearing in missions