Airfield taxi path updates
Add taxi paths for Bodenplatte, and also update the rules in the
airfield test to try to avoid some taxi issues which have been reported.
Taxi paths in existing maps updated to comply with these rules.
Also move the fake airfield location to the parking location when using
cold starts, as this avoids a weird issue with planes not moving at
mission start.